Refactor index.py, add dataframes to psql database
