+++ gitignore
@@ -1,0 +1,18 @@
+.env
+__pycache__
+.coverage
+htmlcov
+*.pyc
+*.pyo
+__init__.py
+pip-log.txt
+pip-delete-this-directory.txt
+.tox/
+.cache
+nosetests.xml
+
+basic_mvp_data/
+*.xlsx
+*.csv
+
+*.sav
